refactor(server): extract admin-ack correlation + favorites services (#3962 4.2a PR4)
Move pendingAdminAcks/sendAdminCommand(AwaitAck)/tryResolveAdminAck (now
resolveByRequestId)/settleAdminAck into a new adminTransactionService, and
supportsFavorites/sendFavoriteNode(AwaitAck)/sendRemoveFavoriteNode/
checkAutoFavorite/autoFavoriteSweep into a new favoritesService, per
task42a_spec.md §4d/§4c. Manager keeps thin public delegates for the
externally-consumed API (routes/automation) and private delegates for the
two internal-only auto-favorite entry points.

favoritesSupportCache and autoFavoritingNodes stay physically on the manager
(bridged via narrow accessors) rather than moving, because
favoritesSupport.test.ts, passiveMode.test.ts, and
autoFavorite.perSource.test.ts all poke those fields directly and must stay
green unchanged; autoFavoriteSweepRunning has no such pin and moved fully.
FavoritesService is constructed with AdminTransactionService injected
directly (not routed through the manager's delegate) since it's a pure
request/response dependency with no other manager-level side effects.

logOutgoingPacket/parseFirmwareVersion/localNodeSettingKey were widened from
private to (default) public, matching the sendWantConfigId/
replaceAnnouncementTokens precedent from PR2/PR3, since they're shared with
other unmoved manager code.
